# Make use of faster RNG implementation for ionization