**Important refactoring**

Greatly overdue, the random number generator for the ADK model (and all future rate models for ionization) has been exchanged. 
Before we used an RNG that was creating new states during each time step of the simulation which were unique for each step, particle species, cell and MPI rank.
Now the state is created only once and then saved as well as changed for each cell after each use. This implies the uniqueness of the seed for the aforementioned conditions and reduces workload at the same time.

This now allows for faster testing of all the ionization models and more importantly: **highly increased simulation speed** for more sophisticated ionization models than BSI.

Big kudos to @Flamefire ! :tada: 

This pull request depends on the changes in #1541