@n01r 2016-08-02
*Very (!)* naive runtime testing of the branch: 

Ions activated, Ionization with `ADKLinPol`, 1 node, 4 NVIDIA K20 GPUs
`mpiexec -n 4 ./picongpu -d 2 2 1 -g 128 128 64 -s 500 --e_png.period 32 --e_png.axis yz --e_png.slicePoint 0.5 --e_png.folder pngElectronsYZ`

**before**
```
initialization time: 58sec 521msec = 58 sec
  0 % =        0 | time elapsed:                    0msec | avg time per step:   0msec
  5 % =       25 | time elapsed:             6sec 187msec | avg time per step: 247msec
 10 % =       50 | time elapsed:            12sec 371msec | avg time per step: 247msec
 15 % =       75 | time elapsed:            18sec 557msec | avg time per step: 247msec
 20 % =      100 | time elapsed:            24sec 742msec | avg time per step: 247msec
 25 % =      125 | time elapsed:            30sec 928msec | avg time per step: 247msec
 30 % =      150 | time elapsed:            37sec 111msec | avg time per step: 247msec
 35 % =      175 | time elapsed:            43sec 340msec | avg time per step: 249msec
 40 % =      200 | time elapsed:            49sec 798msec | avg time per step: 258msec
 45 % =      225 | time elapsed:            56sec 501msec | avg time per step: 268msec
 50 % =      250 | time elapsed:       1min  3sec 376msec | avg time per step: 274msec
 55 % =      275 | time elapsed:       1min 10sec 374msec | avg time per step: 279msec
 60 % =      300 | time elapsed:       1min 17sec 515msec | avg time per step: 285msec
 65 % =      325 | time elapsed:       1min 24sec 797msec | avg time per step: 291msec
 70 % =      350 | time elapsed:       1min 32sec 265msec | avg time per step: 298msec
 75 % =      375 | time elapsed:       1min 39sec 761msec | avg time per step: 299msec
 80 % =      400 | time elapsed:       1min 47sec 160msec | avg time per step: 295msec
 85 % =      425 | time elapsed:       1min 54sec 450msec | avg time per step: 291msec
 90 % =      450 | time elapsed:       2min  1sec 644msec | avg time per step: 287msec
 95 % =      475 | time elapsed:       2min  8sec 771msec | avg time per step: 285msec
100 % =      500 | time elapsed:       2min 15sec 834msec | avg time per step: 282msec
calculation  simulation time:  2min 15sec 834msec = 135 sec
full simulation time:  3min 14sec 759msec = 194 sec
```

**after**
```
initialization time: 58sec 208msec = 58 sec
  0 % =        0 | time elapsed:                    0msec | avg time per step:   0msec
  5 % =       25 | time elapsed:                  441msec | avg time per step:  17msec
 10 % =       50 | time elapsed:                  881msec | avg time per step:  17msec
 15 % =       75 | time elapsed:             1sec 319msec | avg time per step:  17msec
 20 % =      100 | time elapsed:             1sec 756msec | avg time per step:  17msec
 25 % =      125 | time elapsed:             2sec 192msec | avg time per step:  17msec
 30 % =      150 | time elapsed:             2sec 629msec | avg time per step:  17msec
 35 % =      175 | time elapsed:             3sec 113msec | avg time per step:  19msec
 40 % =      200 | time elapsed:             3sec 840msec | avg time per step:  29msec
 45 % =      225 | time elapsed:             4sec 816msec | avg time per step:  39msec
 50 % =      250 | time elapsed:             5sec 959msec | avg time per step:  45msec
 55 % =      275 | time elapsed:             7sec 230msec | avg time per step:  50msec
 60 % =      300 | time elapsed:             8sec 646msec | avg time per step:  56msec
 65 % =      325 | time elapsed:            10sec 194msec | avg time per step:  61msec
 70 % =      350 | time elapsed:            11sec 928msec | avg time per step:  69msec
 75 % =      375 | time elapsed:            13sec 696msec | avg time per step:  70msec
 80 % =      400 | time elapsed:            15sec 364msec | avg time per step:  66msec
 85 % =      425 | time elapsed:            16sec 919msec | avg time per step:  62msec
 90 % =      450 | time elapsed:            18sec 371msec | avg time per step:  58msec
 95 % =      475 | time elapsed:            19sec 755msec | avg time per step:  55msec
100 % =      500 | time elapsed:            21sec  81msec | avg time per step:  53msec
calculation  simulation time: 21sec  81msec = 21 sec
full simulation time:  1min 19sec 679msec = 79 sec
```

**speedup: 6.43x** (ratio of the calculation simulation times)