@bussmann 2016-08-02
Nice! Is this the speedup you expected?