@PrometheusPi 2016-08-09
#1547 has been merged