@PrometheusPi 2016-08-10
@n01r please:
 -  rebase against the current `dev`
 - safeguard your RNG init with a Boolean OR in `MySimulation.hpp